requirements.txt
gspread==6.1.2
pandas==2.2.2
Python-binance==1.0.19
Requests==2.32.3
ta==0.11.0